requirements.txt
asgiref==3.10.0
attrs==25.4.0
black==25.9.0
cfgv==3.4.0
click==8.3.0
distlib==0.4.0
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
Faker==37.11.0
filelock==3.20.0
flake8==7.3.0
identify==2.6.15
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pre_commit==4.3.0
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
pytokens==0.2.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
virtualenv==20.35.3
